Added settings files Added modular formatters
- Added PEP0257 compliant formatter

- Added google style formatter

- Added numpy style formatter

- Added sphinx reST style formatter
